fix: consolidate parallel child tests to satisfy pkg file-count limit
